+++ nn_dataflow/__init__.py
@@ -34,7 +34,7 @@ from .LoopBlockingScheme import LoopBlockingScheme
 from .MapStrategy import MapStrategy, MapStrategyEyeriss
 from .NestedLoopDesc import NestedLoopDesc
 from .Network import Network
-from .NNDataflow import SchedulingResultDict
+from .NNDataflowScheme import NNDataflowScheme
 from .Option import Option
 from .PartitionScheme import PartitionScheme
 from .PhyDim2 import PhyDim2